requirements.txt
dj-database-url==0.4.1
Django==1.9.7
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==2.0.6
httplib2==0.10.3
google-api-python-client==1.6.2
oauth2client==4.1.1
gitpython==2.1.5
requests==2.9.1